MemberController 시큐리티로직 충돌 해결

+++ backend/src/main/java/com/codestates/member/controller/MemberController.java
@@ -1,11 +1,13 @@
 package com.codestates.member.controller;
 
+import com.codestates.auth.utils.ErrorResponse;
 import com.codestates.common.response.MultiResponseDto;
 import com.codestates.member.dto.MemberDto;
 import com.codestates.member.entity.Member;
 import com.codestates.member.entity.MemberRoom;
 import com.codestates.member.mapper.MemberMapper;
 import com.codestates.member.service.MemberService;
+import org.springframework.security.core.Authentication;
 import org.springframework.data.domain.Page;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -15,6 +17,7 @@ import org.springframework.web.bind.annotation.*;
 import javax.validation.Valid;
 import javax.validation.constraints.Positive;
 import java.util.List;
+import java.util.Map;
 
 
 @Validated
@@ -41,9 +44,20 @@ public class MemberController {
 
 
 
+
     @PatchMapping("/{member-id}/edit")
     public ResponseEntity patchMember(@PathVariable("member-id") @Positive long memberId,
-                                      @Valid @RequestBody MemberDto.Patch requestBody) {
+                                      @Valid @RequestBody MemberDto.Patch requestBody,
+                                      Authentication authentication) {
+
+        Map<String, Object> principal = (Map) authentication.getPrincipal();
+        long jwtMemberId = ((Number) principal.get("sub")).longValue();
+
+        if(jwtMemberId != (memberId)) {
+            ErrorResponse errorResponse = ErrorResponse.of(HttpStatus.FORBIDDEN, "토큰 불일치 : 권한이 없는 사용자 입니다.");
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(errorResponse);
+        }
+
         requestBody.setMemberId(memberId);
         Member member = memberService.updateMember(mapper.patchDtoToMember(requestBody));
         return new ResponseEntity<>(mapper.memberToPatchResponseDto(member),HttpStatus.OK);
@@ -51,11 +65,20 @@ public class MemberController {
 
 
 
+
     @GetMapping("/{member-id}")
-    public ResponseEntity getMember(@PathVariable("member-id") @Positive long memberId){
+    public ResponseEntity getMember(@PathVariable("member-id") @Positive long memberId,
+                                    Authentication authentication) {
+
+        if(authentication == null) {
+            ErrorResponse errorResponse = ErrorResponse.of(HttpStatus.UNAUTHORIZED, "미인증 사용자 : 토큰이 존재하지 않습니다.");
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(errorResponse);
+        }
+
         Member member = memberService.findMember(memberId);
         return new ResponseEntity<>(mapper.memberToGetResponseDto(member), HttpStatus.OK);
     }
+
 
 
 
@@ -72,10 +95,20 @@ public class MemberController {
 
 
 
+
     @GetMapping("/{member-id}/like")
-    public ResponseEntity getLikeRooms(@PathVariable("member-id") @Positive long memberId,
-                                       @Positive @RequestParam("page") int page,
-                                       @Positive @RequestParam("size") int size) {
+    public ResponseEntity getLikeRooms(@Positive @RequestParam("page") int page,
+                                       @Positive @RequestParam("size") int size,
+                                       @PathVariable("member-id") @Positive long memberId,
+                                       Authentication authentication) {
+
+        Map<String, Object> principal = (Map) authentication.getPrincipal();
+        long jwtMemberId = ((Number) principal.get("sub")).longValue();
+
+        if(jwtMemberId != (memberId)) {
+            ErrorResponse errorResponse = ErrorResponse.of(HttpStatus.FORBIDDEN, "토큰 불일치 : 권한이 없는 사용자 입니다.");
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(errorResponse);
+        }
 
         Page<MemberRoom> memberRoomPage = memberService.findLikeRooms(page -1, size, memberId);
         List<MemberRoom> memberRoomList = memberRoomPage.getContent();
@@ -87,11 +120,22 @@ public class MemberController {
 
 
 
+
     @GetMapping("/{member-id}/created")
-    public ResponseEntity getCreatedRoom(@PathVariable("member-id") @Positive long memberId,
-                                         @Positive @RequestParam("page") int page,
-                                         @Positive @RequestParam("size") int size) {
-        Page<MemberRoom> memberRoomPage = memberService.findCreatedRooms(page -1, size, memberId);
+    public ResponseEntity getCreatedRoom(@Positive @RequestParam("page") int page,
+                                         @Positive @RequestParam("size") int size,
+                                         @PathVariable("member-id") @Positive long memberId,
+                                         Authentication authentication) {
+
+        Map<String, Object> principal = (Map) authentication.getPrincipal();
+        long jwtMemberId = ((Number) principal.get("sub")).longValue();
+
+        if(jwtMemberId != (memberId)) {
+            ErrorResponse errorResponse = ErrorResponse.of(HttpStatus.FORBIDDEN, "토큰 불일치 : 권한이 없는 사용자 입니다.");
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(errorResponse);
+        }
+
+        Page<MemberRoom> memberRoomPage = memberService.findCreatedRooms(page-1, size, memberId);
         List<MemberRoom> memberRoomList = memberRoomPage.getContent();
         List<MemberDto.CreatedRoomResponseDtos> responseDtosList = mapper.memberToCreatedResponseDtos(memberRoomList);
 
@@ -101,11 +145,22 @@ public class MemberController {
 
 
 
+
     @GetMapping("/{member-id}/record")
-    public ResponseEntity getRecordRoom(@PathVariable("member-id") @Positive long memberId,
-                                        @Positive @RequestParam("page") int page,
-                                        @Positive @RequestParam("size") int size) {
-        Page<MemberRoom> memberRoomPage = memberService.findRecordRooms(page -1, size, memberId);
+    public ResponseEntity getRecordRoom(@Positive @RequestParam("page") int page,
+                                        @Positive @RequestParam("size") int size,
+                                        @PathVariable("member-id") @Positive long memberId,
+                                        Authentication authentication) {
+
+        Map<String, Object> principal = (Map) authentication.getPrincipal();
+        long jwtMemberId = ((Number) principal.get("sub")).longValue();
+
+        if(jwtMemberId != (memberId)) {
+            ErrorResponse errorResponse = ErrorResponse.of(HttpStatus.FORBIDDEN, "토큰 불일치 : 권한이 없는 사용자 입니다.");
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(errorResponse);
+        }
+
+        Page<MemberRoom> memberRoomPage = memberService.findRecordRooms(page-1, size, memberId);
         List<MemberRoom> memberRoomList = memberRoomPage.getContent();
         List<MemberDto.RecordRoomResponseDtos> responseDtosList = mapper.memberToRecordResponseDtos(memberRoomList);
 
@@ -115,8 +170,19 @@ public class MemberController {
 
 
 
+
     @DeleteMapping("/{member-id}")
-    public ResponseEntity deleteMember(@PathVariable("member-id") @Positive long memberId) {
+    public ResponseEntity deleteMember(@PathVariable("member-id") @Positive long memberId,
+                                       Authentication authentication) {
+
+        Map<String, Object> principal = (Map) authentication.getPrincipal();
+        long jwtMemberId = ((Number) principal.get("sub")).longValue();
+
+        if(jwtMemberId != (memberId)) {
+            ErrorResponse errorResponse = ErrorResponse.of(HttpStatus.FORBIDDEN, "토큰 불일치 : 권한이 없는 사용자 입니다.");
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).body(errorResponse);
+        }
+
         memberService.removeUser(memberId);
         return new ResponseEntity<>(HttpStatus.OK);
     }
